+++ R/floating_catchment_area.R
@@ -9,9 +9,7 @@
 #' @template land_use_data
 #' @template opportunity
 #' @template travel_cost
-#' @param demand A string. The name of the column in `land_use_data` with the
-#'   number of people in each origin that will be considered potential
-#'   competitors.
+#' @template demand
 #' @param method A string. Which floating catchment area measure to use.
 #'   Current available options are `"2sfca"` and `"bfca"`. More info in the
 #'   details.